npm start serves precompiled pages with node, npm run dev runs webpack, npm run build works
