+++ server.js
@@ -1,39 +1,21 @@
 /* eslint no-console: 0 */
-import path from 'path';
-import express from 'express';
-import webpack from 'webpack';
-import webpackMiddleware from 'webpack-dev-middleware';
-import webpackHotMiddleware from 'webpack-hot-middleware';
-import config from './webpack.config.js';
+const express = require('express');
+const fs = require('fs');
+const path = require('path');
 
-const isDeveloping = process.env.NODE_ENV !== 'production';
-const port = isDeveloping ? 3735 : process.env.PORT;
 const app = express();
+const port = process.env.PORT || 3735;
 
-const compiler = webpack(config);
-const middleware = webpackMiddleware(compiler, {
-  publicPath: config.output.publicPath,
-  contentBase: 'src',
-  stats: {
-    colors: true,
-    hash: false,
-    timings: true,
-    chunks: false,
-    chunkModules: false,
-    modules: false
-  }
-});
-
-app.use(middleware);
-app.use(webpackHotMiddleware(compiler));
+// serve up compiled static assets if we're in production mode
+app.use(express.static(__dirname + '/dist'));
 app.get('*', function response(req, res) {
-  res.write(middleware.fileSystem.readFileSync(path.join(__dirname, 'dist/index.html')));
-  res.end();
+  res.sendFile(path.join(__dirname, 'dist/index.html'));
 });
 
 app.listen(port, 'localhost', function onStart(err) {
   if (err) {
     console.log(err);
+  } else {
+    console.info('==> 🌎 Listening on port %s. Open up http://localhost:%s/ in your browser.', port, port);
   }
-  console.info('==> 🌎 Listening on port %s. Open up http://localhost:%s/ in your browser.', port, port);
 });
